+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0]
+
+### Fixed
+
+- React Agent with SO
+
+### Added
+
+- Complex input
+
+### Updated
+
+- Dependencies
+- Error handling
+- Tests
+
 ## [0.5.0]
 
 ### Fixed